Move default props to provider
In an attempt to share defaults and make it less confusing which props
are coming from the provider and which props are getting passed into the
component, this moves all default props to the provider and creates more
type safety for the date picker components that get data from this
provider.

A potential downside of this change is that it makes it more difficult
to use the SingleDatePicker component and RangeDatePicker in isolation
if that is something we want to do later.