🏗🐛 Fix compiling JSON Schema in ES5 (#37974)
Fixes the following `esbuild` error:

```
error: Transforming default arguments to the configured target environment ("es5") is not supported yet
```

This would occur due to this part of the transform:

```diff
  function validate(
    data,
-   {
-     instancePath = '',
-     foo: 'bar',
-   }
+   instancePath = ''
  ) {}
```

When replacing the `ObjectPattern` with `instancePath = ''`, we were previously inserting the `ObjectProperty` rather than its child `AssignmentPattern`.

This would create a tree that _looked_ right when generated, but which Babel did not how to transpile into ES5. Tests aren't updated with this change because the string output is exactly the same.