🏗 Tree-shake included JSON Schema in compiled output (#37973)
ajv's output may sometimes reference a part of the schema object, in which case the entire schema is included in the bundle.

This change follows the references made to each schema object, in order to only preserve the portions required. This is a big culprit of compiled size, since the schema varies in length and may include a lot of unnecessary information.