Except Google Translate from the font-stylesheet timeout
We're seeing a huge number of CSP violations because of Chrome's translate features injects this stylesheet dynamically. Somehow, they've privileged it to run regardless of our strict CSP. But when we try to force the first render of the page by toggling it's `media` attribute, the CSP fails and the stylesheet is prevented from loading.