Explicitly cast a char32_t to uint32_t before streaming it to output.
In C++17, the char32_t is silently formatted as a number, but in C++20, overloads of << accepting wide character types are deleted, breaking compilation of this code.

The value also was formatted in decimal afaict, but all other values here are formatted in hex, so do that here as well.

PiperOrigin-RevId: 499523425