cleans up JsonDict and JsonArray
JsonDict and JsonArray form members to a vector of an incomplete type, which is
not allowed. libc++ has historically turned a blind eye here, but an incoming
patch will cause these types to break. This commit intends to get in front of
that libc++ change by moving all function definitions that reference `items_`
to after `JsonObject` is defined.

There were also a few other opportunities for reducing complexity that I snuck
into this CL, such as correcting use of forward references and ensuring
`types.h` and `types.cc` include what they use.

PiperOrigin-RevId: 653431283